use exec if there's just 1 command; otherwise check for availability of pkill, ps, etc

+++ templates/estafette-entrypoint.sh
@@ -2,10 +2,22 @@
 
 set -e
 
+{{- if .UseExec }}
+{{range .Commands}}
+exec {{.}}
+{{end}}
+{{- else }}
+
 forward_sigterm() {
     echo "Received SIGTERM, forwarding to child processes..."
     trap - 15
-    pkill -P $$
+    if [ -x "$(command -v pkill)" ]; then
+        pkill -P $$
+    elif [ -x "$(command -v ps)" ]; then
+        ps -o pid --ppid $$ --no-heading | xargs kill
+    else
+        echo "Sorry, this container doesn't seem to have the commands (pkill, ps, xargs) to reliably kill child processes"
+    fi
 }
 
 trap forward_sigterm 15
@@ -13,4 +25,5 @@ trap forward_sigterm 15
 {{range .Commands}}
     {{- .}} &
     wait
+{{end}}
 {{end}}
